fix: core editor related code cleanup
