+++ packages/react-tinacms-editor/src/core/styles/CodeMirror.tsx
