+++ packages/react-tinacms-editor/src/core/styles/ProseMirror.tsx
