+++ packages/react-tinacms-editor/src/state/plugins/Menu/historyControl.tsx
@@ -19,18 +19,15 @@ limitations under the License.
 import * as React from 'react'
 
 import { RedoIcon, UndoIcon } from '@tinacms/icons'
-import { EditorView } from 'prosemirror-view'
 import { undo, redo, undoDepth, redoDepth } from 'prosemirror-history'
 
 import { MenuButton } from './MenuComponents'
+import { useEditorStateContext } from '../../../core/context/editorState'
 
-interface HistoryProps {
-  view: EditorView
-}
-
-export const UndoControl = ({ view }: HistoryProps) => {
+export const UndoControl = () => {
+  const { editorView } = useEditorStateContext()
   const undoChange = () => {
-    const { state, dispatch } = view
+    const { state, dispatch } = editorView!.view
     undo(state, dispatch)
   }
 
@@ -39,16 +36,18 @@ export const UndoControl = ({ view }: HistoryProps) => {
       data-tooltip="Undo"
       data-side="top"
       onClick={undoChange}
-      disabled={undoDepth(view.state) < 1}
+      disabled={undoDepth(editorView!.view.state) < 1}
     >
       <UndoIcon />
     </MenuButton>
   )
 }
 
-export const RedoControl = ({ view }: HistoryProps) => {
+export const RedoControl = () => {
+  const { editorView } = useEditorStateContext()
+
   const redoChange = () => {
-    const { state, dispatch } = view
+    const { state, dispatch } = editorView!.view
     redo(state, dispatch)
   }
 
@@ -57,7 +56,7 @@ export const RedoControl = ({ view }: HistoryProps) => {
       data-tooltip="Redo"
       data-side="top"
       onClick={redoChange}
-      disabled={redoDepth(view.state) < 1}
+      disabled={redoDepth(editorView!.view.state) < 1}
     >
       <RedoIcon />
     </MenuButton>
