+++ packages/react-tinacms-editor/src/state/plugins/Menu/markControl.tsx
@@ -20,10 +20,7 @@ import * as React from 'react'
 import { EditorView } from 'prosemirror-view'
 import { toggleMark } from 'prosemirror-commands'
 import { MenuButton } from './MenuComponents'
-
-export interface Props {
-  view: EditorView
-}
+import { useEditorStateContext } from '../../../core/context/editorState'
 
 export interface Options {
   mark: string
@@ -47,85 +44,79 @@ export function markControl({
   isDisabled,
   onClick,
 }: Options) {
-  return class _ extends React.Component<Props, any> {
-    static displayName = `${mark}Control`
+  return () => {
+    const { editorView } = useEditorStateContext()
+    const view = editorView!.view
 
-    markType(markName: string) {
-      const schema = this.props.view.state.schema
+    const markType = (markName: string) => {
+      const schema = view.state.schema
       return schema.marks[markName]
     }
 
-    get active(): boolean {
-      return this.markIsActive(mark)
-    }
+    const isActive = () => markIsActive(mark)
 
-    markIsActive = (markName: string): boolean => {
-      const { state } = this.props.view
-      const mark = this.markType(markName)
+    const markIsActive = (markName: string): boolean => {
+      const { state } = view
+      const mark = markType(markName)
       const { from, $from, to, empty } = state.selection
       if (empty) return !!mark.isInSet(state.storedMarks || $from.marks())
       else return state.doc.rangeHasMark(from, to, mark)
     }
 
-    get disabled(): boolean {
-      if (isDisabled) return isDisabled(this.props.view)
+    const isOptionDisabled = () => {
+      if (isDisabled) return isDisabled(view)
       if (mark === 'image')
         if (selectionOnly) {
-          const { $cursor } = this.props.view.state.selection as any
-          return (
-            !!$cursor || this.inCodeBlock || this.incompatibleMarksAreActive
-          )
+          const { $cursor } = view.state.selection as any
+          return !!$cursor || isInCodeBlock() || isIncompatibleMarksAreActive()
         }
 
-      return this.inCodeBlock || this.incompatibleMarksAreActive
+      return isInCodeBlock() || isIncompatibleMarksAreActive()
     }
 
-    get inCodeBlock(): boolean {
-      const view = this.props.view
+    const isInCodeBlock = () => {
       const node = view.state.selection.$from.node(
         view.state.selection.$from.depth
       )
       return node.type === view.state.schema.nodes.code_block
     }
 
-    get incompatibleMarksAreActive(): boolean {
+    const isIncompatibleMarksAreActive = () => {
       return noMix
-        .map(this.markIsActive)
+        .map(markIsActive)
         .reduce(
           (someMarkActive, nextMarkActive) => nextMarkActive || someMarkActive,
           false
         )
     }
 
-    onClick = () => {
+    const onOptionClick = () => {
       if (onClick) {
-        onClick(this.props.view)
+        onClick(view)
       }
-      if (this.disabled) return
-      const { state, dispatch } = this.props.view
+      if (isOptionDisabled()) return
+      const { state, dispatch } = view
 
       if ((state.selection as any).$cursor && selectionOnly) {
         return
       }
 
-      toggleMark(this.markType(mark), defaultAttrs)(state, dispatch)
+      toggleMark(markType(mark), defaultAttrs)(state, dispatch)
     }
 
-    render() {
-      if (!this.markType(mark)) {
-        return null
-      }
-      return (
-        <MenuButton
-          data-tooltip={tooltip}
-          data-side="top"
-          onClick={this.onClick}
-          active={!this.disabled && this.active}
-          disabled={this.disabled}
-        >
-          <Icon />
-        </MenuButton>
-      )
+    if (!markType(mark)) {
+      return null
     }
+    return (
+      <MenuButton
+        data-tooltip={tooltip}
+        data-side="top"
+        onClick={onOptionClick}
+        active={!isOptionDisabled() && isActive()}
+        disabled={isOptionDisabled()}
+      >
+        <Icon />
+      </MenuButton>
+    )
   }
 }
